Backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
fastapi==0.110.0
h11==0.14.0
idna==3.4
motor==3.3.1
pydantic==2.6.3
pydantic-core==2.16.3
pymongo==4.5.0
sniffio==1.3.0
starlette==0.36.3
typing-extensions==4.8.0
uvicorn==0.28.0
fastapi-socketio==0.0.10
python-socketio ==5.11.3
python-engineio == 4.9.1
netifaces >= 0.10.6
tomli==2.0.1
websockets==10.4